+++ src/redux/actions/actionTypes.js
@@ -1,14 +1,3 @@
 export const actionTypes = {
-    GET_USERS: 'recipes/getUsers',
-    GET_USER: 'recipes/getUser',
-    DELETE_USER: 'recipes/deleteUser',
-    EDIT_USER: 'recipes/editUser',
-    DELETEALL_USERS: 'recipes/deleteAllUser',
-    ADD_USER: 'recipes/addUser',
-    GET_QUESTIONS: 'questions/getQuestions',
-    GET_QUESTION: 'questions/getQuestion',
-    GENERATE_RECOMMENDATION: 'questions/generateRecommendation',
-    GET_DESTINATIONS: 'questions/getDestinations',
-    GET_DESTINATION: 'questions/getDestination',
-    LOGIN_USER: 'recipes/loginUser',
+    LOGIN_USER: 'user/loginUser',
 };
