--- src/redux/actions/index.js
@@ -1,13 +1,0 @@
-export const addCard = (cardObject) => {
-	return {
-		type: 'ADD_CARD',
-		payload: cardObject
-	}
-}
-
-export const deleteCard = (id) => {
-	return {
-		type: 'DELETE_CARD',
-		payload: id
-	}
-}
